+++ plugins/source/gcp/codegen/main.go
@@ -70,6 +70,83 @@ func needsProjectIDColumn(r recipes.Resource) bool {
 	return r.Multiplex != &recipes.OrgMultiplex
 }
 
+func InferListFunction(r *recipes.Resource) {
+	// We can infer the List function by matching a list function that return the same struct as r.Struct
+	server := reflect.TypeOf(r.RegisterServer).In(1)
+	for i := 0; i < server.NumMethod(); i++ {
+		method := server.Method(i)
+		if strings.HasPrefix(method.Name, "List") {
+			response := method.Type.Out(0).Elem()
+			for j := 0; j < response.NumField(); j++ {
+				field := response.Field(j)
+				kind := field.Type.Kind()
+				if kind != reflect.Slice {
+					continue
+				}
+				if field.Type.Elem() == reflect.TypeOf(r.Struct) {
+					r.ListFunctionName = method.Name
+					r.RequestStructName = method.Type.In(1).Elem().Name()
+					r.ResponseStructName = method.Type.Out(0).Elem().Name()
+					return
+				}
+			}
+		}
+	}
+}
+
+func getColumn(columns codegen.ColumnDefinitions, name string) *codegen.ColumnDefinition {
+	for i := range columns {
+		if columns[i].Name == name {
+			return &columns[i]
+		}
+	}
+	return nil
+}
+
+func listMethodSignature(method reflect.Method) string {
+	t := method.Type
+	buf := strings.Builder{}
+	buf.WriteString(method.Name + "(")
+	buf.WriteString(t.In(0).String())
+	buf.WriteString(", ")
+	buf.WriteString(t.In(1).String())
+	buf.WriteString(")")
+	buf.WriteString(" (")
+	buf.WriteString(t.Out(0).String())
+	buf.WriteString(", ")
+	buf.WriteString(t.Out(1).String())
+	buf.WriteString(")")
+	return buf.String()
+}
+
+func isValidListMethod(method reflect.Method) bool {
+	if !strings.HasPrefix(method.Name, "List") {
+		return false
+	}
+
+	_, responseHasNextPage := method.Type.Out(0).Elem().FieldByName("NextPageToken")
+	return responseHasNextPage
+}
+
+func generateMockTestData(r *recipes.Resource) {
+	registerServerPath := runtime.FuncForPC(reflect.ValueOf(r.RelationsTestData.RegisterServer).Pointer()).Name()
+	serverName := strings.Split(registerServerPath, "/")[4]
+	pbName := strings.Split(serverName, ".")[0]
+	pbPath := strings.Split(registerServerPath, pbName)[0]
+	r.RelationsTestData.ProtobufImport = fmt.Sprintf("%s%s", pbPath, pbName)
+	r.RelationsTestData.RegisterServerName = serverName
+	r.RelationsTestData.UnimplementedServerName = strings.Replace(serverName, "Register", "Unimplemented", 1)
+	server := reflect.TypeOf(r.RelationsTestData.RegisterServer).In(1)
+	for i := 0; i < server.NumMethod(); i++ {
+		method := server.Method(i)
+		if isValidListMethod(method) {
+			sig := listMethodSignature(method)
+			responseType := method.Type.Out(0).Elem().String()
+			r.RelationsTestData.ListFunctions = append(r.RelationsTestData.ListFunctions, recipes.ListFunctions{Signature: sig, ResponseStructName: responseType})
+		}
+	}
+}
+
 func generateResource(r recipes.Resource, mock bool) {
 	var err error
 	_, filename, _, ok := runtime.Caller(0)
@@ -82,10 +159,12 @@ func generateResource(r recipes.Resource, mock bool) {
 		path := strings.Split(runtime.FuncForPC(reflect.ValueOf(r.NewFunction).Pointer()).Name(), ".")
 		r.NewFunctionName = path[len(path)-1]
 	}
+
 	if r.RegisterServer != nil {
 		path := strings.Split(runtime.FuncForPC(reflect.ValueOf(r.RegisterServer).Pointer()).Name(), ".")
 		r.RegisterServerName = path[len(path)-1]
 		r.UnimplementedServerName = strings.Replace(r.RegisterServerName, "Register", "Unimplemented", 1)
+
 	}
 
 	if r.ClientName == "" && r.NewFunctionName != "" {
@@ -97,23 +176,30 @@ func generateResource(r recipes.Resource, mock bool) {
 		r.StructName = reflect.TypeOf(r.Struct).Elem().Name()
 	}
 
-	if r.ListFunction != nil && !r.SkipFetch {
-		path := strings.Split(runtime.FuncForPC(reflect.ValueOf(r.ListFunction).Pointer()).Name(), ".")
-		r.ListFunctionName = path[len(path)-1]
-		// https://stackoverflow.com/questions/32925344/why-is-there-a-fm-suffix-when-getting-a-functions-name-in-go
-		r.ListFunctionName = strings.Split(r.ListFunctionName, "-")[0]
-		r.RequestStructName = reflect.TypeOf(r.ListFunction).In(1).Elem().Name()
-
-		switch {
-		case r.RegisterServer != nil:
-			server := reflect.TypeOf(r.RegisterServer).In(1)
-			method, _ := server.MethodByName(r.ListFunctionName)
-			r.ResponseStructName = method.Type.Out(0).Elem().Name()
-		case r.ListFunctionName == "Get":
-			r.ResponseStructName = r.StructName
-		default:
-			r.ResponseStructName = r.StructName + r.ListFunctionName
+	if !r.SkipFetch {
+		if r.ListFunction == nil {
+			InferListFunction(&r)
+		} else {
+			path := strings.Split(runtime.FuncForPC(reflect.ValueOf(r.ListFunction).Pointer()).Name(), ".")
+			r.ListFunctionName = path[len(path)-1]
+			// https://stackoverflow.com/questions/32925344/why-is-there-a-fm-suffix-when-getting-a-functions-name-in-go
+			r.ListFunctionName = strings.Split(r.ListFunctionName, "-")[0]
+			r.RequestStructName = reflect.TypeOf(r.ListFunction).In(1).Elem().Name()
+			switch {
+			case r.RegisterServer != nil:
+				server := reflect.TypeOf(r.RegisterServer).In(1)
+				method, _ := server.MethodByName(r.ListFunctionName)
+				r.ResponseStructName = method.Type.Out(0).Elem().Name()
+			case r.ListFunctionName == "Get":
+				r.ResponseStructName = r.StructName
+			default:
+				r.ResponseStructName = r.StructName + r.ListFunctionName
+			}
 		}
+	}
+
+	if mock && r.RelationsTestData.RegisterServer != nil {
+		generateMockTestData(&r)
 	}
 
 	if r.ResponseStruct != nil {
@@ -191,10 +277,16 @@ func generateResource(r recipes.Resource, mock bool) {
 	}
 
 	for _, f := range r.PrimaryKeys {
-		for i := range r.Table.Columns {
-			if r.Table.Columns[i].Name == f {
-				r.Table.Columns[i].Options.PrimaryKey = true
-			}
+		column := getColumn(r.Table.Columns, f)
+		if column != nil {
+			column.Options.PrimaryKey = true
+		}
+	}
+
+	for _, f := range r.IgnoreInTestsColumns {
+		column := getColumn(r.Table.Columns, f)
+		if column != nil {
+			column.IgnoreInTests = true
 		}
 	}