+++ plugins/source/snyk/codegen/resources/transformers.go
@@ -1,0 +1,21 @@
+package resources
+
+import (
+	"reflect"
+
+	"github.com/cloudquery/plugin-sdk/caser"
+)
+
+var csr = caser.New(
+	caser.WithCustomInitialisms(
+		map[string]bool{
+			"ACL":  true,
+			"DERP": true,
+			"SSH":  true,
+		},
+	),
+)
+
+func nameTransformer(f reflect.StructField) (string, error) {
+	return csr.ToSnake(f.Name), nil
+}